fix(exports): rm redundant exports, keep Routing and Method only.
