+++ express-zod-api/tests/index.spec.ts
@@ -1,24 +1,5 @@
-import type { IRouter } from "express";
-import { z } from "zod";
 import * as entrypoint from "../src";
-import type {
-  ApiResponse,
-  AppConfig,
-  BasicSecurity,
-  BearerSecurity,
-  CommonConfig,
-  CookieSecurity,
-  FlatObject,
-  HeaderSecurity,
-  IOSchema,
-  InputSecurity,
-  LoggerOverrides,
-  Method,
-  OAuth2Security,
-  OpenIdSecurity,
-  Routing,
-  ServerConfig,
-} from "../src";
+import type { Routing, LoggerOverrides, Method } from "../src";
 
 describe("Index Entrypoint", () => {
   describe("exports", () => {
@@ -33,48 +14,10 @@ describe("Index Entrypoint", () => {
       if (entity !== undefined) expect(entity).toMatchSnapshot();
     });
 
-    test("Issue 952, 1182, 1269: should expose certain types and interfaces", () => {
+    test("should expose certain types and interfaces", () => {
       expectTypeOf<"get">().toExtend<Method>();
-      expectTypeOf(z.object({})).toExtend<IOSchema>();
-      expectTypeOf({}).toExtend<FlatObject>();
       expectTypeOf({}).toEqualTypeOf<LoggerOverrides>();
       expectTypeOf({}).toExtend<Routing>();
-      expectTypeOf<{
-        cors: true;
-        logger: { level: "silent" };
-      }>().toExtend<CommonConfig>();
-      expectTypeOf<{
-        app: IRouter;
-        cors: true;
-        logger: { level: "silent" };
-      }>().toExtend<AppConfig>();
-      expectTypeOf<{
-        http: { listen: 1234 };
-        logger: { level: "silent" };
-        cors: false;
-      }>().toExtend<ServerConfig>();
-      expectTypeOf<{ type: "basic" }>().toEqualTypeOf<BasicSecurity>();
-      expectTypeOf<{
-        type: "bearer";
-        format?: string;
-      }>().toEqualTypeOf<BearerSecurity>();
-      expectTypeOf<{
-        type: "cookie";
-        name: string;
-      }>().toEqualTypeOf<CookieSecurity>();
-      expectTypeOf<{
-        type: "header";
-        name: string;
-      }>().toEqualTypeOf<HeaderSecurity>();
-      expectTypeOf<{ type: "input"; name: string }>().toEqualTypeOf<
-        InputSecurity<string>
-      >();
-      expectTypeOf<{ type: "oauth2" }>().toExtend<OAuth2Security<string>>();
-      expectTypeOf<{
-        type: "openid";
-        url: string;
-      }>().toEqualTypeOf<OpenIdSecurity>();
-      expectTypeOf({ schema: z.string() }).toExtend<ApiResponse<z.ZodString>>();
     });
   });
 });
