+++ express-zod-api/tests/testing.spec.ts
@@ -5,8 +5,8 @@ import {
   ResultHandler,
   testEndpoint,
   testMiddleware,
-  type CommonConfig,
 } from "../src";
+import type { CommonConfig } from "../src/config-type";
 import type { Mock } from "vitest";
 
 describe("Testing", () => {
